Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 int main() {
-    float a,b,c;
+    int a,b,c;
     printf("Average: ");
-    scanf("%f",&a);
-    scanf("%f",&b);
-    scanf("%f",&c);
+    scanf("%d",&a);
+    scanf("%d",&b);
+    scanf("%d",&c);
     printf("%.2f",a+b+c/3);
     return 0;
 }
